requirements.txt
python-magic-bin==0.4.14
exifread==3.0.0
mutagen==1.47.0
hachoir==3.2.0
PyPDF2==3.0.1
python-docx==1.1.0
tinytag==1.10.1
Pillow>=10.0.0
